requirements.txt
allure-pytest==2.13.5
pytest==8.3.2
requests==2.32.3
selene==2.0.0rc9
curlify==2.2.1